feat: add styles and text to sign-up page
