+++ src/App.tsx
@@ -5,7 +5,7 @@ import BoardsPage from 'pages/BoardsPage/BoardsPage';
 import HomePage from 'pages/HomePage/HomePage';
 import LoginPage from 'pages/LoginPage';
 import NotfoundPage from 'pages/NotFoundPage/NotFoundPage';
-import SignUpPage from 'pages/SignUpPage';
+import SignUpPage from 'pages/SignUpPage/SignUpPage';
 
 import Layout from 'components/Layout/Layout';
