+++ src/pages/SignUpPage/SignUpPage.module.scss
@@ -1,0 +1,89 @@
+@import url('../../variables/Variables.module.scss');
+
+.signup {
+  background: var(--color-darkpurple);
+}
+
+.container {
+  padding: 40px 0;
+}
+
+.mainContainer {
+  display: flex;
+  justify-content: space-between;
+  min-height: calc(100vh - 145px);
+  gap: 40px;
+}
+
+.inputData {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 40px;
+
+  input {
+    color: var(--color-darkpurple);
+    font-size: 18px;
+    font-family: 'Montserrat-Regular';
+    padding: 15px;
+    padding-left: 25px;
+    border-radius: 20px;
+    border: 1px solid var(--color-black);
+    -webkit-box-shadow: 0px 0px 20px 13px rgba(0, 0, 0, 0.5);
+    -moz-box-shadow: 0px 0px 20px 13px rgba(0, 0, 0, 0.5);
+    box-shadow: 0px 0px 20px 13px rgba(0, 0, 0, 0.5);
+  }
+}
+
+.title {
+  font-size: 40px;
+  color: var(--color-red);
+  font-family: 'Montserrat-Bold';
+}
+
+.imageBackground {
+  display: flex;
+  width: 50%;
+  flex-direction: column;
+  align-items: center;
+  gap: 20px;
+  border-radius: 20px;
+  padding: 40px;
+  background: var(--color-red);
+  box-sizing: border-box;
+
+  img {
+    width: 90%;
+    object-fit: contain;
+    margin-bottom: 20px;
+  }
+}
+
+.imageTitleContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.imageTitle {
+  font-size: 40px;
+  color: var(--color-white);
+  font-family: 'Montserrat-Bold';
+  text-transform: uppercase;
+}
+
+.imageSubTitle {
+  font-size: 22px;
+  color: var(--color-white);
+  filter: brightness(80%);
+  font-family: 'Montserrat-Regular';
+}
+
+.signUpBtn {
+  font-size: 24px;
+  color: var(--color-white);
+  font-family: 'Montserrat-Regular';
+  padding: 15px 50px;
+  background: var(--color-red);
+  border-radius: 20px;
+}
